LayoutTests/platform/chromium/inspector/timeline/timeline-script-tag-1-expected.txt

CONSOLE MESSAGE: line 8: iframe loaded
Tests the Timeline API instrumentation of an HTML script tag.


ParseHTML
----> ScheduleStyleRecalculation
----> InvalidateLayout
ParseHTML
----> ScheduleStyleRecalculation
----> EvaluateScript
--------> TimeStamp : SCRIPT TAG
EvaluateScript Properties:
{
    children : <object>
    counters : <object>
    data : {
        lineNumber : <number>
        url : <string>
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "EvaluateScript"
    usedHeapSize : <number>
}
ParseHTML
----> InvalidateLayout
ParseHTML
----> InvalidateLayout
----> EventDispatch
--------> FunctionCall
------------> FunctionCall
----> MarkLoad
----> Layout
----> MarkDOMContent
